+++ pystac/extensions/single_file_stac.py
@@ -1,179 +1,137 @@
-import json
 from pystac.catalog import Catalog
 
-from pystac import STACError
+import pystac
+from pystac import (STACError, Extensions)
 from pystac.collection import Collection
-from pystac.stac_io import STAC_IO
-from pystac.item import Item
+from pystac.extensions.base import (CatalogExtension, ExtensionDefinition, ExtendedObject)
 
 
-class SingleFileSTAC(Catalog):
-    """Provides a set of Collections and Items as a single file catalog. The single file
+def create_single_file_stac(catalog):
+    """Creates a Single File STAC from a STAC catalog.
+
+    This method will recursively collect any collections and items in the catalog
+    and return a new catalog with the same properties as the old one, with cleared
+    links and the 'collections' and 'features' property of the Single File STAC holding
+    each of the respective collected STAC objects.
+
+    Collections will be filtered to only those referenced by items via the collection_id.
+    All links in the items and collections will be cleared in the Single File STAC.
+
+    Args:
+        catalog (Catalog): Catalog to walk while constructin the Single File STAC
+    """
+    collections = {}
+    items = []
+    for root, _, cat_items in catalog.walk():
+        if issubclass(type(root), Collection):
+            new_collection = root.clone()
+            new_collection.clear_links()
+            collections[root.id] = new_collection
+        for item in cat_items:
+            new_item = item.clone()
+            new_item.clear_links()
+            items.append(new_item)
+
+    filtered_collections = []
+    for item in items:
+        if item.collection_id in collections:
+            filtered_collections.append(collections[item.collection_id])
+            collections.pop(item.collection_id)
+
+    result = catalog.clone()
+    result.clear_links()
+    result.ext.enable(Extensions.SINGLE_FILE_STAC)
+    result.ext[Extensions.SINGLE_FILE_STAC].apply(features=items, collections=filtered_collections)
+
+    return result
+
+
+class SingleFileSTACCatalogExt(CatalogExtension):
+    """An extension of Catalog that provides a set of Collections and Items
+    as a single file catalog. A SingleFileSTAC
     is a self contained catalog that contains everything that would normally be in a
     linked set of STAC files.
 
     Args:
-        id (str): Identifier for the Catalog. Must be unique within the STAC.
-        description (str): Detailed multi-line description to fully explain the catalog.
-            `CommonMark 0.28 syntax <http://commonmark.org/>`_ MAY be used for rich text
-            representation.
-        features (List[Item]): Optional initial list of items contained by
-            this SingleFileSTAC.
-        stac_extensions (List[str]): Optional list of extensions implemented.
-        collections (List[Collection]): Optional initial list of collections contained
-            by this SingleFileSTAC.
-        search (Search): Optional search information associated with this SingleFileSTAC.
+        catalog (Catalog): The catalog to be extended.
 
     Attributes:
-        id (str): Identifier for the catalog.
-        description (str): Detailed multi-line description to fully explain the catalog.
-        features (List[Item]): Items contained by this SingleFileSTAC
-        collections (List[Collection]): Collections contained
-            by this SingleFileSTAC.
-        search (Search): Optional search information associated with this SingleFileSTAC.
-        links (List[Link]): A list of :class:`~pystac.Link` objects representing
-            all links associated with this SingleFileSTAC.
+        catalog (Catalog): The catalog that is being extended.
 
+    Note:
+        Using SingleFileSTACCatalogExt to directly wrap a Catalog will
+        add the 'proj' extension ID to the catalog's stac_extensions.
     """
-    def __init__(self,
-                 id,
-                 description,
-                 features=None,
-                 stac_extensions=None,
-                 collections=None,
-                 search=None):
-        super().__init__(id, description, title=None, stac_extensions=stac_extensions)
+    def __init__(self, catalog):
+        if catalog.stac_extensions is None:
+            catalog.stac_extensions = [Extensions.SINGLE_FILE_STAC]
+        elif Extensions.SINGLE_FILE_STAC not in catalog.stac_extensions:
+            catalog.stac_extensions.append(Extensions.SINGLE_FILE_STAC)
 
+        self.catalog = catalog
+
+    def apply(self, features, collections=None):
+        """
+        Args:
+            features (List[Item]): List of items contained by
+                this SingleFileSTAC.
+            collections (List[Collection]): Optional list of collections that are
+                used by any of teh Items in the catalog.
+        """
+        self.features = features
+        self.collections = collections
+
+    @classmethod
+    def enable_extension(cls, catalog):
+        # Ensure the 'type' property is correct so that the Catalog is valid GeoJSON.
+        catalog.extra_fields['type'] = 'FeatureCollection'
+
+    @property
+    def features(self):
+        """Get or sets a list of :class:`~pystac.Item` contained in this Single File STAC.
+
+        Returns:
+            List[Item]
+        """
+        features = self.catalog.extra_fields.get('features')
         if features is None:
-            self.features = []
+            raise STACError('Invalid Single File STAC: does not have "features" property.')
+
+        return [pystac.read_dict(feature) for feature in features]
+
+    @features.setter
+    def features(self, v):
+        self.catalog.extra_fields['features'] = [item.to_dict() for item in v]
+
+    @property
+    def collections(self):
+        """Get or sets a list of :class:`~pystac.Collection` objects contained
+        in this Single File STAC.
+
+        Returns:
+            List[Band]
+        """
+        collections = self.catalog.extra_fields.get('collections')
+
+        if collections is not None:
+            collections = [pystac.read_dict(col) for col in collections]
+        return collections
+
+    @collections.setter
+    def collections(self, v):
+        if v is not None:
+            self.catalog.extra_fields['collections'] = [col.to_dict() for col in v]
         else:
-            self.features = features
+            self.catalog.extra_fields.pop('collections', None)
 
-        if collections is None:
-            self.collections = []
-        else:
-            self.collections = collections
+    @classmethod
+    def _object_links(cls):
+        return []
 
-        self.search = search
-
-    def __repr__(self):
-        return '<SingleFileSTAC with {} items and {} collections>'.format(
-            len(self.features), len(self.collections))
-
-    def get_collections(self):
-        """Gets all the collections associated with this SingleFileSTAC.
-
-        Returns:
-            List[Collection]: The Collections of this SingleFileSTAC
-        """
-        return self.collections
-
-    def add_collection(self, collection):
-        """Adds a Collection to this SingleFileSTAC.
-
-        Args:
-            collection (Collection): The collection to add.
-        """
-        self.collections.append(collection)
-
-    def add_collections(self, collections):
-        """Adds Collections to this SingleFileSTAC.
-
-        Args:
-            collections (Iterable[Collection]): The collections to add.
-        """
-        for collection in collections:
-            self.add_collection(collection)
-
-    @staticmethod
-    def from_file(uri):
-        """Reads an SingleFileSTAC from a file.
-
-        Args:
-            href (str): The HREF to read the item from.
-
-        Returns:
-            SingleFileSTAC: SingleFileSTAC that was read from the given file.
-        """
-
-        d = json.loads(STAC_IO.read_text(uri))
-        c = SingleFileSTAC.from_dict(d)
-        return c
-
-    @staticmethod
-    def from_dict(d):
-        """Constructs an SingleFileSTAC from a dict.
-
-        Returns:
-            SingleFileSTAC: The SingleFileSTAC deserialized from the JSON dict.
-        """
-        features = [Item.from_dict(feature) for feature in d['features']]
-        collections = [Collection.from_dict(c) for c in d['collections']]
-        stac_extensions = d.get('stac_extensions')
-
-        # Tie together items to their collections
-        collection_dict = dict([(c.id, c) for c in collections])
-        for item in features:
-            if item.collection_id is not None:
-                if item.collection_id not in collection_dict:
-                    raise STACError('Collection with id {} is referenced '
-                                    'by item {}, but is not in the collections '
-                                    'of this SingleFileSTAC'.format(item.collection_id, item.id))
-                item.set_collection(collection_dict[item.collection_id])
-
-        search_obj = None
-        if 'search' in d.keys():
-            sd = d['search']
-            search_obj = Search(sd.get('endpoint'), sd.get('parameters'))
-        return SingleFileSTAC(features, stac_extensions, collections, search_obj)
-
-    def to_dict(self, include_self_link=False):
-        """Generate a dictionary representing the JSON of this SingleFileSTAC.
-
-        Args:
-            include_self_link (bool): If True, writes the SingleFileSTAC's self link.
-                Defaults to False.
-
-        Returns:
-            dict: A serializion of the SingleFileSTAC that can be written out as JSON.
-        """
-
-        d = super().to_dict(include_self_link=include_self_link)
-
-        d['collections'] = [c.to_dict() for c in self.collections]
-        d['features'] = [f.to_dict() for f in self.features]
-        if self.search:
-            d['search'] = self.search.to_dict()
-
-        return d
+    @classmethod
+    def from_catalog(cls, catalog):
+        return SingleFileSTACCatalogExt(catalog)
 
 
-class Search:
-    """Contains search endpoint and parameters information for a search used to generate
-    a Single File STAC.
-
-    Args:
-        endpoint (str): The root endpoint of a STAC API used for this search.
-        parameters (dict): A dictionary of all the parameters used for the search.
-
-    Attributes:
-        endpoint (str): The root endpoint of a STAC API used for this search.
-        parameters (dict): A dictionary of all the parameters used for the search.
-    """
-    def __init__(self, endpoint=None, parameters=None):
-        self.endpoint = endpoint
-        self.parameters = parameters
-
-    @staticmethod
-    def from_dict(d):
-        return Search(d.get('endpoint'), d.get('parameters'))
-
-    def to_dict(self):
-        d = {}
-        if self.endpoint is not None:
-            d['endpoint'] = self.endpoint
-
-        if self.parameters is not None:
-            d['parameters'] = self.parameters
-
-        return d
+SFS_EXTENSION_DEFINITION = ExtensionDefinition(Extensions.SINGLE_FILE_STAC,
+                                               [ExtendedObject(Catalog, SingleFileSTACCatalogExt)])